add ssher (pronounced "SSH-er") module to builderdash project
Needs further cleanup and a test-suite.

The ssher library uses the Python paramiko library and implements a python class, called "SSHConnection", which enables the caller to connect and authenticate with a remote SSH server either directly, or via an ssh bastion host (aka proxy aka jump-host). Once the underlying paramiko SSHClient is successfully connected and authenticated, a SFTP connection is also established to the target host so that file systems operations on the target may be issued via the client.

The library implements a configurable retry with delay feature for establishing connections to servers.

The library implements an ssh_run_command which executes a blocking shell command a remote host. The exit code of the command is returned to the call along with optionally returning the stdout and/or stderr bytes generated by the command as well. As the remote command executes and the client receives buffered stdout and/or stderr streams, that function may also optionally log those outputs using log functions such as log.info and/or log.error so that output for long-running commands may be displayed as it is received.