enable specification of a custom k8s StorageClass for kubevirt builds
You may now specify a k8s StorageClass via parameter "kubevirt_storage_class_name", and build.py instructs builderdash to use the value for kubevirt builds. The StorageClass specified must exist in the cluster. null / "None" values will cause builderdash to omit the specification of a custom StorageClass and the build will therefor use the default StorageClass. Custom StorageClasses enable us to leverage features like SSD-backed disks and specify a custom StorageClass 'reclaimPolicy' such as 'Retain' instead of the default 'Delete'. All of this is an effort to use faster storage for builds and enable the disks to outlive the k8s cluster in case we want to reuse the builds in a new k8s cluster / not leave a cluster running.